+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # dev
 
-- change1
+- adding cli options, set spec as global
+- made the poc simpler